templates,xmleval: split the interpreter/CreateUIObject blob into a new module
templates.lua's job is the template registry (GetTemplateOrThrow/
SetTemplate/GetTemplateInfo) -- that was its correct, minimal scope
before it grew to also hold CreateUIObject and the whole XML interpreter
in the previous commit, just to keep them co-located with the registry
lookups they use.

Move that blob into a new xmleval module instead, and have templates.lua
re-export CreateUIObject/frames/LoadFile from it. This keeps the registry
under its own name, gives the interpreter+object-creation code an
accurate name of its own, and needs zero changes to any consumer --
api.lua, loader.lua, gencode.lua, cgencode.lua, mainloop.lua, runner.lua,
and the data/uiobjects and data/impl luafile snippets all still say
templates.CreateUIObject/templates.LoadFile/etc, unchanged from the
previous commit.

The dependency edge only runs one way (templates depends on xmleval);
xmleval doesn't depend back on templates since it kept its own registry
implementation rather than delegating back, which is what avoids
recreating the cycle this whole reorg exists to break.
